organizacion tablas

+++ db/migrate/20130723151619_create_cotizaciondetalles.rb
@@ -2,8 +2,8 @@ class CreateCotizaciondetalles < ActiveRecord::Migration
   def change
     create_table :cotizaciondetalles do |t|
       t.string :producto_id
-      t.float :cantidad
-      t.float :subtotal
+      t.integer :cantidad
+      t.integer :subtotal
       t.integer :cotizacion_id
 
       t.timestamps
